carefirst/pyproject.toml
[tool.poetry]
name = "carefirst"
version = "0.1.0"
description = ""
authors = ["Bronte Pendergast <bronpend@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
pandas = "^2.2.0"
pypdf = "^4.0.1"
sentence-transformers = "^2.3.1"
faiss-cpu = "^1.7.4"
gradio = "^4.16.0"
pydantic = "^2.6.1"
black = "^24.1.1"
ruff = "^0.2.1"
isort = "^5.13.2"
fastapi = "^0.109.2"
requests = "^2.31"
fastapi-cache2 = {extras = ["redis"], version = "^0.1.9"}
uvicorn = "^0.27.0.post1"
langchain = "^0.1.4"
langchain-core = ">=0.1.0,<0.2.0"
#langchain-community = "^0.0.19"
pymongo = "^4.6.1"
python-dotenv = "^1.0.1"
openai = ">=1.10.0,<2.0.0"
nemoguardrails = "^0.7.1"
pydantic-mongo = "^2.1.0"
langchain-openai = "^0.0.7"

[tool.poetry.group.dev.dependencies]
nltk = "^3.8.1"
spacy = "^3.7.2"
gensim = "^4.3.2"
wordcloud = "^1.9.3"
matplotlib-venn = "^0.11.10"
matplotlib-venn-wordcloud = "^0.2.6"
requests = "^2.31.0"
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
